refactor: split host autostart routes
